backend/requirements.txt
django
djangorestframework
djoser
django-cors-headers
